# Perf: Optimize collections and enumerators