--- a/src/Microsoft.OData.Core/InstanceAnnotationWriteTracker.cs
+++ b/src/Microsoft.OData.Core/InstanceAnnotationWriteTracker.cs
@@ -18,14 +18,13 @@ namespace Microsoft.OData
         /// Maintains the write status for each annotation using its key.
         /// If a key exists in the list then it is considered written.
         /// </summary>
-        private readonly HashSet<string> writeStatus;
+        private HashSet<string> writeStatus;
 
         /// <summary>
         /// Creates a new <see cref="InstanceAnnotationWriteTracker"/> to hold write status for instance annotations.
         /// </summary>
         public InstanceAnnotationWriteTracker()
         {
-            this.writeStatus = new HashSet<string>(StringComparer.Ordinal);
         }
 
         /// <summary>
@@ -35,7 +34,7 @@ namespace Microsoft.OData
         /// <param name="key">The key of the element to check if its written.</param>
         public bool IsAnnotationWritten(string key)
         {
-            return this.writeStatus.Contains(key);
+            return this.writeStatus != null && this.writeStatus.Contains(key);
         }
 
         /// <summary>
@@ -45,6 +44,11 @@ namespace Microsoft.OData
         /// <param name="key">The key of the element to mark as written.</param>
         public bool MarkAnnotationWritten(string key)
         {
+            if (this.writeStatus == null)
+            {
+                this.writeStatus = new HashSet<string>(StringComparer.Ordinal);
+            }
+
             return this.writeStatus.Add(key);
         }
     }