… src/Microsoft.OData.Core/Json/ODataJsonWriterUtils.cs
… src/Microsoft.OData.Core/ODataAnnotatable.cs
… src/Microsoft.OData.Core/ODataWriterCore.cs
… test/FunctionalTests/Microsoft.OData.Core.Tests/Json/JsonLightInstanceAnnotationWriterTests.cs
… test/FunctionalTests/Microsoft.OData.Core.Tests/Json/ODataJsonWriterUtilsAsyncTests.cs
… test/PerformanceTests/ComponentTests/ODataWriter/ODataWriterFeedTests.cs