<!-- markdownlint-disable MD002 MD041 -->

### Issues

*This pull request fixes #2190.*

### Description

*Briefly describe the changes of this pull request.*
- [x] Eliminate `IEnumerator` heap allocations in `WriteInstanceAnnotations` when called with a `List<T>` argument:
  - initialize `ODataAnnotatable` `instanceAnnotations` with a `List<T>` instead of `Collection<T>`
  - cast annotations `IEnumerable` to a `List<T>` if possible when running `foreach` loop
- [x] Reduce `List<InstanceAnnotation>` allocations where it's not used
  - lazy initialization of the `instanceAnnotations` collection in `ODataAnnotatable`
- [x] Eliminate `HashSet<string>` heap allocations when instance annotations collection is empty
  - Lazy initialization of the `writeStatus` hash set in `InstanceAnnotationWriterTracker`
- [x] Eliminate enumerators and iterators allocations from accessing ScopeStack's parent and grandparent:
  - Implement `ScopeStack` using `List<Scope>` instead of `Stack<Scope>` so we can access parents cheaply using index
- [x] Replicate optimizations in async versions where applicable

The optimizations in `WriteInstanceAnnotations` have 2 versions of a foreach loop, one against a `List<T>` instance and the other against the `IEnumerable<T>` instance. When performing the foreach loop against a `List<T>`, the `List<T>.Enumerator` is not allocated on the heap because it's a struct. However, when running a foreach loop against the same collection through an `IEnumerable<T>` interface, the enumerator will be accessed through an `IEnumerator<T>` reference which causes it to be boxed and therefore copied to the heap. There are a lot of such allocations that would add up to noticeable improvements if we optimized them, but I admit that duplicating the foreach loop isn't ideal. And since some of these properties are exposed in public APIs as `IEnumerable<T>` or `ICollection<T>`, I can't just hardcode them to `List<T>`. That's why I did not go ahead and apply these optimizations to more areas of the code. The for loops in the `WriteIntsanceAnnotations` method account for 0.29% of allocations AGS, that's why I thought it would still be worth going ahead with it. If I addressed all such allocations, it would probably save 1+% in AGS, but I wanted to get feedback on whether this duplicate foreach-loop is worth the price.

There was a suggestion from @g2mula to create helper methods that encapsulate this "double" for-each loop, something like `collection.ForEach(action)` extension method that tries to cast the collection to a list before running the loop. However for this approach to be effective, I'd have two create multiple overloads for different versions to allow the action to accept multiple arguments without leading to closure allocations from captured variables (in same cases I may need to pass as much as 6 variables) (e.g. `ForEach<T>(this IEnumerable<T>, Action<T>)`, `ForEach<T, TArg1>(this IEnumerable<T>, Action<T, TArg1>)`, `ForEach<T, TArg1, TArg2>(this IEnumerable<T>, Action<TArg1>, Action<TArg2>)`, etc.). I would also need to account for async versions of those methods. And in some cases the loop needs to terminate prematurely when a condition is met. This seemed to be an explosion of complexity that seems to be overkill, it made very hesitant to commit it to this PR.

### Benchmarks

There's seems to be a 15-25% reduction in allocated memory size across the benchmarks as a result of the changes.

| Slower                                                                           | diff/base | Base Median (ns) | Diff Median (ns) | Modality|
| -------------------------------------------------------------------------------- | ---------:| ----------------:| ----------------:| -------- |
| SerializationBaselineTests.SerializationBenchmarks.WriteJson(dataSize: 5000, isM |      1.05 |       8898300.00 |       9318600.00 |         |
| SerializationBaselineTests.SerializationBenchmarks.WriteJson(dataSize: 5000, isM |      1.04 |       8811100.00 |       9146850.00 |         |
| SerializationBaselineTests.SerializationBenchmarks.WriteJson(dataSize: 10000, is |      1.03 |      17782200.00 |      18353600.00 |         |
| SerializationBaselineTests.SerializationBenchmarks.WriteJson(dataSize: 10000, is |      1.03 |      17808000.00 |      18339200.00 |         |
| SerializationBaselineTests.SerializationBenchmarks.WriteJson(dataSize: 1000, isM |      1.02 |       1920100.00 |       1958800.00 | several?|

| Faster                                                                           | base/diff | Base Median (ns) | Diff Median (ns) | Modality|
| -------------------------------------------------------------------------------- | ---------:| ----------------:| ----------------:| --------:|
| SerializationBaselineTests.SerializationBenchmarks.WriteODataSyncFile(dataSize:  |      1.24 |      25225200.00 |      20343300.00 |         |
| SerializationBaselineTests.SerializationBenchmarks.WriteODataSync(dataSize: 1000 |      1.24 |      23862600.00 |      19262100.00 |         |
| SerializationBaselineTests.SerializationBenchmarks.WriteODataSync(dataSize: 5000 |      1.23 |     118341100.00 |      95905100.00 |         |
| SerializationBaselineTests.SerializationBenchmarks.WriteODataSync(dataSize: 1000 |      1.23 |      25351900.00 |      20552100.00 |         |
| SerializationBaselineTests.SerializationBenchmarks.WriteODataSyncArrayPool(dataS |      1.23 |      23740000.00 |      19270400.00 |         |
| SerializationBaselineTests.SerializationBenchmarks.WriteODataSync(dataSize: 1000 |      1.23 |     237077400.00 |     192992900.00 |         |
| SerializationBaselineTests.SerializationBenchmarks.WriteODataSync(dataSize: 1000 |      1.23 |     252622700.00 |     205743500.00 |         |
| SerializationBaselineTests.SerializationBenchmarks.WriteODataSyncFile(dataSize:  |      1.23 |     123499700.00 |     100619500.00 |         |
| SerializationBaselineTests.SerializationBenchmarks.WriteODataSyncArrayPool(dataS |      1.23 |     118660300.00 |      96805100.00 |         |
| SerializationBaselineTests.SerializationBenchmarks.WriteODataSyncFile(dataSize:  |      1.23 |     260084050.00 |     212241800.00 |         |
| SerializationBaselineTests.SerializationBenchmarks.WriteODataSyncArrayPool(dataS |      1.22 |     237068500.00 |     193569850.00 |         |
| SerializationBaselineTests.SerializationBenchmarks.WriteODataSyncArrayPool(dataS |      1.22 |     126436600.00 |     103321700.00 |         |
| SerializationBaselineTests.SerializationBenchmarks.WriteODataSyncFile(dataSize:  |      1.22 |     243478000.00 |     199052750.00 |         |
| SerializationBaselineTests.SerializationBenchmarks.WriteODataSyncArrayPoolFile(d |      1.22 |     260673500.00 |     213368700.00 |         |
| SerializationBaselineTests.SerializationBenchmarks.WriteODataSyncArrayPoolFile(d |      1.22 |     123509100.00 |     101132600.00 |         |
| SerializationBaselineTests.SerializationBenchmarks.WriteODataSync(dataSize: 5000 |      1.22 |     125474200.00 |     102864000.00 |         |
| SerializationBaselineTests.SerializationBenchmarks.WriteODataSyncArrayPoolFile(d |      1.22 |     129935600.00 |     106530300.00 |         |
| SerializationBaselineTests.SerializationBenchmarks.WriteODataSyncArrayPoolFile(d |      1.22 |      24971900.00 |      20498700.00 |         |
| SerializationBaselineTests.SerializationBenchmarks.WriteODataSyncArrayPool(dataS |      1.22 |     251591850.00 |     206530700.00 |         |
| SerializationBaselineTests.SerializationBenchmarks.WriteODataSyncArrayPoolFile(d |      1.22 |     243142250.00 |     199607300.00 |         |
| SerializationBaselineTests.SerializationBenchmarks.WriteODataSyncFile(dataSize:  |      1.22 |      26332700.00 |      21657100.00 |         |
| SerializationBaselineTests.SerializationBenchmarks.WriteODataSyncFile(dataSize:  |      1.21 |     129795700.00 |     107022400.00 |         |
| SerializationBaselineTests.SerializationBenchmarks.WriteODataSyncArrayPool(dataS |      1.21 |      25103000.00 |      20721100.00 |         |
| SerializationBaselineTests.SerializationBenchmarks.WriteODataSyncArrayPoolFile(d |      1.20 |      26290200.00 |      21824850.00 |         |
| SerializationBaselineTests.SerializationBenchmarks.WriteODataFile(dataSize: 1000 |      1.09 |      81137800.00 |      74461350.00 |         |
| SerializationBaselineTests.SerializationBenchmarks.WriteOData(dataSize: 1000, is |      1.09 |      81192600.00 |      74658700.00 |         |
| SerializationBaselineTests.SerializationBenchmarks.WriteODataArrayPoolFile(dataS |      1.09 |      81098100.00 |      74643500.00 |         |
| SerializationBaselineTests.SerializationBenchmarks.WriteODataArrayPool(dataSize: |      1.09 |      79381500.00 |      73063700.00 |         |
| SerializationBaselineTests.SerializationBenchmarks.WriteOData(dataSize: 1000, is |      1.09 |      79374950.00 |      73058300.00 |         |
| SerializationBaselineTests.SerializationBenchmarks.WriteODataFile(dataSize: 5000 |      1.08 |     406913950.00 |     375129800.00 |         |
| SerializationBaselineTests.SerializationBenchmarks.WriteODataArrayPool(dataSize: |      1.08 |      81239100.00 |      74901000.00 |         |
| SerializationBaselineTests.SerializationBenchmarks.WriteODataFile(dataSize: 1000 |      1.08 |     813155600.00 |     750092800.00 |         |
| SerializationBaselineTests.SerializationBenchmarks.WriteODataFile(dataSize: 1000 |      1.08 |     792749100.00 |     732295100.00 |         |
| SerializationBaselineTests.SerializationBenchmarks.WriteODataFile(dataSize: 5000 |      1.08 |     396905100.00 |     366955100.00 |         |
| SerializationBaselineTests.SerializationBenchmarks.WriteOData(dataSize: 10000, i |      1.08 |     809794600.00 |     748916450.00 |         |
| SerializationBaselineTests.SerializationBenchmarks.WriteODataArrayPoolFile(dataS |      1.08 |     792409500.00 |     733425600.00 |         |
| SerializationBaselineTests.SerializationBenchmarks.WriteODataArrayPool(dataSize: |      1.08 |     808588600.00 |     748517400.00 |         |
| SerializationBaselineTests.SerializationBenchmarks.WriteOData(dataSize: 10000, i |      1.08 |     790945750.00 |     732221800.00 |         |
| SerializationBaselineTests.SerializationBenchmarks.WriteODataArrayPoolFile(dataS |      1.08 |     396616900.00 |     367378500.00 |         |
| SerializationBaselineTests.SerializationBenchmarks.WriteODataArrayPoolFile(dataS |      1.08 |     810447100.00 |     751194400.00 |         |
| SerializationBaselineTests.SerializationBenchmarks.WriteOData(dataSize: 5000, is |      1.08 |     394974000.00 |     366193000.00 |         |
| SerializationBaselineTests.SerializationBenchmarks.WriteODataArrayPool(dataSize: |      1.08 |     394269300.00 |     365613000.00 |         |
| SerializationBaselineTests.SerializationBenchmarks.WriteODataArrayPoolFile(dataS |      1.08 |     405924650.00 |     376516100.00 |         |
| SerializationBaselineTests.SerializationBenchmarks.WriteODataArrayPool(dataSize: |      1.08 |     789774950.00 |     732646500.00 |         |
| SerializationBaselineTests.SerializationBenchmarks.WriteODataArrayPool(dataSize: |      1.07 |     404467400.00 |     376915400.00 |         |
| SerializationBaselineTests.SerializationBenchmarks.WriteOData(dataSize: 5000, is |      1.07 |     404288000.00 |     376802400.00 |         |
| SerializationBaselineTests.SerializationBenchmarks.WriteODataFile(dataSize: 1000 |      1.07 |      82057300.00 |      76606000.00 |         |
| SerializationBaselineTests.SerializationBenchmarks.WriteODataArrayPoolFile(dataS |      1.07 |      82417800.00 |      77090850.00 |         |
| SerializationBaselineTests.SerializationBenchmarks.WriteJsonFile(dataSize: 1000, |      1.02 |       2372650.00 |       2330900.00 |         |

The column headers have been truncate because the tables were too long, but here they are:
- Method
- Category
- dataSize
- isModelImmutable
- Mean
- Error
- StdDev
- Median
- Ratio
- RatioSD
- Gen 0
- Gen 1
- Allocated

**Before**

![image](https://user-images.githubusercontent.com/8460169/137278294-16eb6f2f-da7c-42dc-955c-562efd4d5c2e.png)

**After**

![image](https://user-images.githubusercontent.com/8460169/137278356-5228521b-92f1-4e2d-9e46-433c7c3d9466.png)

**Before**

![image](https://user-images.githubusercontent.com/8460169/137701428-b43d3b3a-819f-4264-a9a1-34fc0742c66c.png)


**After**

![image](https://user-images.githubusercontent.com/8460169/137701284-0aa53d6a-6dab-46e6-9fce-baf21cc22e92.png)


### Checklist (Uncheck if it is not completed)

- [x] *Test cases added*
- [ ] *Build and test with one-click build and test script passed*

### Additional work necessary

*If documentation update is needed, please add "Docs Needed" label to the issue and provide details about the required document change in the issue.*